+++ framework/test/mocha/common/application.js
@@ -19,9 +19,9 @@ const rewire = require('rewire');
 const async = require('async');
 const _ = require('lodash');
 const { registeredTransactions } = require('./registered_transactions');
-const ed = require('../../../src/modules/chain/helpers/ed');
-const jobsQueue = require('../../../src/modules/chain/helpers/jobs_queue');
-const Sequence = require('../../../src/modules/chain/helpers/sequence');
+const ed = require('../../../src/modules/chain/utils/ed');
+const jobsQueue = require('../../../src/modules/chain/utils/jobs_queue');
+const Sequence = require('../../../src/modules/chain/utils/sequence');
 const { BlockSlots } = require('../../../src/modules/chain/blocks/block_slots');
 const { createCacheComponent } = require('../../../src/components/cache');
 const { StorageSandbox } = require('./storage_sandbox');
