+++ framework/test/mocha/common/modules_loader.js
@@ -17,11 +17,11 @@
 const express = require('express');
 const randomstring = require('randomstring');
 const async = require('async');
-const Sequence = require('../../../src/modules/chain/helpers/sequence');
+const Sequence = require('../../../src/modules/chain/utils/sequence');
 const { createLoggerComponent } = require('../../../src/components/logger');
 const { ZSchema } = require('../../../src/controller/validator');
-const ed = require('../../../src/modules/chain/helpers/ed');
-const jobsQueue = require('../../../src/modules/chain/helpers/jobs_queue');
+const ed = require('../../../src/modules/chain/utils/ed');
+const jobsQueue = require('../../../src/modules/chain/utils/jobs_queue');
 const Account = require('../../../src/modules/chain/rounds/account');
 
 // TODO: Remove this file
