+++ framework/test/mocha/unit/modules/chain/logic/test_data/dapp.js
@@ -15,7 +15,7 @@
 'use strict';
 
 const crypto = require('crypto');
-const ed = require('../../../../../../../src/modules/chain/helpers/ed');
+const ed = require('../../../../../../../src/modules/chain/utils/ed');
 
 const validPassphrase =
 	'robust weapon course unknown head trial pencil latin acid';
